refactor(providers): share LangChain embedding adapter
Use LangChain embedding clients behind one LlamaIndex-compatible adapter.
